requirements.txt
Django
djangorestframework
psycopg2-binary
pytest
pytest-django
celery
openpyxl
redis
uwsgi
gunicorn